+++ deribit/src/index.ts
@@ -1,6 +1,6 @@
 import { expose, util } from '@chainlink/ea-bootstrap'
 import { makeConfig, makeExecute } from './adapter'
 
-const NAME = 'XBTO'
+const NAME = 'Deribit'
 
 export = { NAME, makeExecute, makeConfig, ...expose(util.wrapExecute(makeExecute())) }